Reorganize tests by dependency tier
