+++ tests/base/test_analysis.py
