+++ tests/base/test_collection.py
