+++ tests/base/test_load_cached_queries.py
