+++ tests/base/test_prebuilt_index.py
