+++ tests/base/test_query_iterator.py
