+++ tests/base/test_search.py
@@ -53,7 +53,7 @@ class TestSearch(unittest.TestCase):
         # Create index without document vectors
         # The current directory depends on if you're running inside an IDE or from command line.
         curdir = os.getcwd()
-        if curdir.endswith('core'):
+        if curdir.endswith('base'):
             corpus_path = '../resources/sample_collection_json'
         else:
             corpus_path = 'tests/resources/sample_collection_json'
