+++ tests/base/test_trectools.py
@@ -26,7 +26,7 @@ class TestTrecTools(unittest.TestCase):
     def setUp(self):
         # The current directory depends on if you're running inside an IDE or from command line.
         curdir = os.getcwd()
-        if curdir.endswith('core'):
+        if curdir.endswith('base'):
             self.root = '../../'
         else:
             self.root = '.'
