+++ tests/core/test_hybrid_search.py
